Extract inquirer from installer
